refactor: enable IS_INCIDENTS_PANEL_V2 feature flag and remove it
