+++ operate/client/src/App/ProcessInstance/IncidentsWrapper/IncidentsFilter/index.test.tsx
@@ -6,80 +6,30 @@
  * except in compliance with the Camunda License 1.0.
  */
 
+import {
+  availableErrorTypes,
+  getIncidentErrorName,
+} from 'modules/utils/incidents';
 import {IncidentsFilter} from './index';
 import {render, screen} from 'modules/testing-library';
 
-import {mockIncidents} from './index.setup';
-import {incidentsStore} from 'modules/stores/incidents';
-import {mockFetchProcessInstanceIncidents} from 'modules/mocks/api/processInstances/fetchProcessInstanceIncidents';
-import {useEffect} from 'react';
-import {ProcessDefinitionKeyContext} from 'App/Processes/ListView/processDefinitionKeyContext';
-import {QueryClientProvider} from '@tanstack/react-query';
-import {getMockQueryClient} from 'modules/react-query/mockQueryClient';
-import {mockFetchProcessDefinitionXml} from 'modules/mocks/api/v2/processDefinitions/fetchProcessDefinitionXml';
-import {IS_INCIDENTS_PANEL_V2} from 'modules/feature-flags';
-
-const {reset, fetchIncidents} = incidentsStore;
-
-const Wrapper = ({children}: {children?: React.ReactNode}) => {
-  useEffect(() => {
-    return reset;
-  }, []);
-
-  return (
-    <ProcessDefinitionKeyContext.Provider value="123">
-      <QueryClientProvider client={getMockQueryClient()}>
-        {children}
-      </QueryClientProvider>
-    </ProcessDefinitionKeyContext.Provider>
-  );
-};
-
 describe('IncidentsFilter', () => {
   it('should render filters', async () => {
-    mockFetchProcessInstanceIncidents().withSuccess(mockIncidents);
-    mockFetchProcessDefinitionXml().withSuccess('');
-
-    await fetchIncidents('1');
-
-    const {user} = render(<IncidentsFilter />, {
-      wrapper: Wrapper,
-    });
+    const {user} = render(<IncidentsFilter />);
 
     await user.click(
       screen.getByRole('combobox', {name: /filter by incident type/i}),
     );
-    expect(
-      screen.getByRole('option', {name: 'Condition error.'}),
-    ).toBeInTheDocument();
-    expect(
-      screen.getByRole('option', {name: 'Extract value error.'}),
-    ).toBeInTheDocument();
 
-    if (IS_INCIDENTS_PANEL_V2) {
-      // TODO: V2 has no flow node filter. Remove code below with full v2 migration in issue https://github.com/camunda/camunda/issues/39548
-      return;
+    for (const errorType of availableErrorTypes) {
+      expect(
+        screen.getByRole('option', {name: getIncidentErrorName(errorType)}),
+      ).toBeInTheDocument();
     }
-    await user.click(
-      screen.getByRole('combobox', {name: /filter by flow node/i}),
-    );
-    expect(
-      screen.getByRole('option', {name: 'flowNodeId_exclusiveGateway'}),
-    ).toBeInTheDocument();
-    expect(
-      screen.getByRole('option', {name: 'flowNodeId_alwaysFailingTask'}),
-    ).toBeInTheDocument();
   });
 
   it('should disable/enable clear all button depending on selected options', async () => {
-    mockFetchProcessInstanceIncidents().withSuccess(mockIncidents);
-    mockFetchProcessDefinitionXml().withSuccess('');
-
-    await fetchIncidents('1');
-
-    const {user} = render(<IncidentsFilter />, {
-      wrapper: Wrapper,
-    });
+    const {user} = render(<IncidentsFilter />);
 
     expect(screen.getByRole('button', {name: 'Reset Filters'})).toBeDisabled();
