+++ operate/client/src/App/ProcessInstance/IncidentsWrapper/IncidentsTable/tests/index.test.tsx
@@ -6,161 +6,172 @@
  * except in compliance with the Camunda License 1.0.
  */
 
-import {IncidentsTable} from '../index';
+import {IncidentsTable} from '..';
 import {formatDate} from 'modules/utils/date';
 import {render, screen, within} from 'modules/testing-library';
-import {incidentsStore} from 'modules/stores/incidents';
 import {Wrapper, incidentsMock, firstIncident, secondIncident} from './mocks';
-import {mockFetchProcessDefinitionXml} from 'modules/mocks/api/v2/processDefinitions/fetchProcessDefinitionXml';
+import {mockFetchProcessInstance} from 'modules/mocks/api/processInstances/fetchProcessInstance';
+import {mockFetchProcessInstance as mockFetchProcessInstanceV2} from 'modules/mocks/api/v2/processInstances/fetchProcessInstance';
+import {
+  createInstance,
+  createProcessInstance,
+  createUser,
+} from 'modules/testUtils';
 import {mockMe} from 'modules/mocks/api/v2/me';
-import {createUser} from 'modules/testUtils';
+import {getIncidentErrorName} from 'modules/utils/incidents';
+
+const firstIncidentErrorName = getIncidentErrorName(firstIncident.errorType);
+const secondIncidentErrorName = getIncidentErrorName(secondIncident.errorType);
 
 describe('IncidentsTable', () => {
-  it('should render the right column headers', () => {
-    mockFetchProcessDefinitionXml().withSuccess('');
-    incidentsStore.setIncidents(incidentsMock);
-
-    render(<IncidentsTable />, {wrapper: Wrapper});
-
-    expect(screen.getByText('Incident Type')).toBeInTheDocument();
-    expect(screen.getByText('Failing Flow Node')).toBeInTheDocument();
-    expect(screen.getByText('Job Id')).toBeInTheDocument();
-    expect(screen.getByText('Creation Date')).toBeInTheDocument();
-    expect(screen.getByText('Error Message')).toBeInTheDocument();
-    expect(screen.getByText('Operations')).toBeInTheDocument();
-    expect(screen.getByText('Root Cause Instance')).toBeInTheDocument();
+  beforeEach(() => {
+    mockFetchProcessInstance().withSuccess(
+      createInstance({permissions: ['UPDATE_PROCESS_INSTANCE']}),
+    );
+    mockFetchProcessInstanceV2().withSuccess(
+      createProcessInstance({
+        hasIncident: true,
+      }),
+    );
   });
 
-  it('should render the right column headers for restricted user', () => {
-    mockFetchProcessDefinitionXml().withSuccess('');
-    incidentsStore.setIncidents(incidentsMock);
-    mockMe().withSuccess(createUser());
+  it('should render incident details', async () => {
+    render(
+      <IncidentsTable
+        state="content"
+        processInstanceKey="1"
+        incidents={incidentsMock}
+      />,
+      {wrapper: Wrapper},
+    );
+    let withinRow = within(
+      screen.getByRole('row', {name: new RegExp(firstIncidentErrorName)}),
+    );
 
-    render(<IncidentsTable />, {wrapper: Wrapper});
+    expect(withinRow.getByText(firstIncidentErrorName)).toBeInTheDocument();
+    expect(withinRow.getByText(firstIncident.elementName)).toBeInTheDocument();
+    expect(withinRow.getByText(firstIncident.jobKey!)).toBeInTheDocument();
+    expect(
+      withinRow.getByText(formatDate(firstIncident.creationTime) || '--'),
+    ).toBeInTheDocument();
+    expect(withinRow.getByText(firstIncident.errorMessage)).toBeInTheDocument();
+    expect(
+      withinRow.queryByRole('button', {name: 'Retry Incident'}),
+    ).not.toBeInTheDocument();
+    withinRow = within(
+      screen.getByRole('row', {name: new RegExp(secondIncidentErrorName)}),
+    );
+    expect(withinRow.getByText(secondIncidentErrorName)).toBeInTheDocument();
+    expect(withinRow.getByText(secondIncident.elementName)).toBeInTheDocument();
+    expect(withinRow.getByText(secondIncident.jobKey!)).toBeInTheDocument();
+    expect(
+      withinRow.getByText(formatDate(secondIncident.creationTime) || '--'),
+    ).toBeInTheDocument();
+    expect(
+      withinRow.getByText(secondIncident.errorMessage),
+    ).toBeInTheDocument();
+    expect(
+      await withinRow.findByRole('button', {name: 'Retry Incident'}),
+    ).toBeInTheDocument();
+  });
+
+  it('should render the right column headers', async () => {
+    render(
+      <IncidentsTable
+        state="content"
+        processInstanceKey="1"
+        incidents={incidentsMock}
+      />,
+      {wrapper: Wrapper},
+    );
 
     expect(screen.getByText('Incident Type')).toBeInTheDocument();
-    expect(screen.getByText('Failing Flow Node')).toBeInTheDocument();
+    expect(screen.getByText('Failing Element')).toBeInTheDocument();
     expect(screen.getByText('Job Id')).toBeInTheDocument();
     expect(screen.getByText('Creation Date')).toBeInTheDocument();
     expect(screen.getByText('Error Message')).toBeInTheDocument();
-    expect(screen.getByText('Operations')).toBeInTheDocument();
-    expect(screen.getByText('Root Cause Instance')).toBeInTheDocument();
+    expect(await screen.findByText('Operations')).toBeInTheDocument();
+  });
+
+  it('should render the right column headers for restricted user', async () => {
+    mockMe().withSuccess(createUser());
+
+    render(
+      <IncidentsTable
+        state="content"
+        processInstanceKey="1"
+        incidents={incidentsMock}
+      />,
+      {wrapper: Wrapper},
+    );
+
+    expect(screen.getByText('Incident Type')).toBeInTheDocument();
+    expect(screen.getByText('Failing Element')).toBeInTheDocument();
+    expect(screen.getByText('Job Id')).toBeInTheDocument();
+    expect(screen.getByText('Creation Date')).toBeInTheDocument();
+    expect(screen.getByText('Error Message')).toBeInTheDocument();
+    expect(await screen.findByText('Operations')).toBeInTheDocument();
   });
 
   it('should render the right column headers for restricted user (with resource-based permissions)', () => {
-    mockFetchProcessDefinitionXml().withSuccess('');
     mockMe().withSuccess(createUser());
     vi.stubGlobal('clientConfig', {
       resourcePermissionsEnabled: true,
     });
 
-    incidentsStore.setIncidents(incidentsMock);
-
-    render(<IncidentsTable />, {wrapper: Wrapper});
+    render(
+      <IncidentsTable
+        state="content"
+        processInstanceKey="1"
+        incidents={incidentsMock}
+      />,
+      {wrapper: Wrapper},
+    );
 
     expect(screen.getByText('Incident Type')).toBeInTheDocument();
-    expect(screen.getByText('Failing Flow Node')).toBeInTheDocument();
+    expect(screen.getByText('Failing Element')).toBeInTheDocument();
     expect(screen.getByText('Job Id')).toBeInTheDocument();
     expect(screen.getByText('Creation Date')).toBeInTheDocument();
     expect(screen.getByText('Error Message')).toBeInTheDocument();
     expect(screen.queryByText('Operations')).not.toBeInTheDocument();
-    expect(screen.getByText('Root Cause Instance')).toBeInTheDocument();
-  });
-
-  it('should render incident details', () => {
-    mockFetchProcessDefinitionXml().withSuccess('');
-    incidentsStore.setIncidents(incidentsMock);
-
-    render(<IncidentsTable />, {wrapper: Wrapper});
-    let withinRow = within(
-      screen.getByRole('row', {
-        name: new RegExp(firstIncident!.errorType.name),
-      }),
-    );
-
-    expect(
-      withinRow.getByText(firstIncident!.errorType.name),
-    ).toBeInTheDocument();
-    expect(withinRow.getByText(firstIncident.flowNodeId)).toBeInTheDocument();
-    expect(withinRow.getByText(firstIncident.jobId!)).toBeInTheDocument();
-    expect(
-      withinRow.getByText(formatDate(firstIncident.creationTime) || '--'),
-    ).toBeInTheDocument();
-    expect(withinRow.getByText(firstIncident.errorMessage)).toBeInTheDocument();
-    expect(
-      withinRow.getByRole('link', {
-        description: /view root cause instance/i,
-      }),
-    ).toBeInTheDocument();
-    expect(
-      withinRow.queryByRole('button', {name: 'Retry Incident'}),
-    ).not.toBeInTheDocument();
-    withinRow = within(
-      screen.getByRole('row', {
-        name: new RegExp(secondIncident!.errorType.name),
-      }),
-    );
-    expect(
-      withinRow.getByText(secondIncident.errorType.name),
-    ).toBeInTheDocument();
-    expect(withinRow.getByText(secondIncident.flowNodeId)).toBeInTheDocument();
-    expect(withinRow.getByText(secondIncident.jobId!)).toBeInTheDocument();
-    expect(
-      withinRow.getByText(formatDate(secondIncident.creationTime) || '--'),
-    ).toBeInTheDocument();
-    expect(
-      withinRow.getByText(secondIncident.errorMessage),
-    ).toBeInTheDocument();
-    expect(
-      withinRow.getByRole('button', {name: 'Retry Incident'}),
-    ).toBeInTheDocument();
   });
 
   it('should render incident details (with resource-based permissions enabled)', () => {
-    mockFetchProcessDefinitionXml().withSuccess('');
     mockMe().withSuccess(createUser());
     vi.stubGlobal('clientConfig', {
       resourcePermissionsEnabled: true,
     });
 
-    incidentsStore.setIncidents(incidentsMock);
-
-    render(<IncidentsTable />, {wrapper: Wrapper});
+    render(
+      <IncidentsTable
+        state="content"
+        processInstanceKey="1"
+        incidents={incidentsMock}
+      />,
+      {wrapper: Wrapper},
+    );
     let withinRow = within(
-      screen.getByRole('row', {
-        name: new RegExp(firstIncident.errorType.name),
-      }),
+      screen.getByRole('row', {name: new RegExp(firstIncidentErrorName)}),
     );
 
-    expect(
-      withinRow.getByText(firstIncident.errorType.name),
-    ).toBeInTheDocument();
-    expect(withinRow.getByText(firstIncident.flowNodeId)).toBeInTheDocument();
-    expect(withinRow.getByText(firstIncident.jobId!)).toBeInTheDocument();
+    expect(withinRow.getByText(firstIncidentErrorName)).toBeInTheDocument();
+    expect(withinRow.getByText(firstIncident.elementName)).toBeInTheDocument();
+    expect(withinRow.getByText(firstIncident.jobKey!)).toBeInTheDocument();
     expect(
       withinRow.getByText(formatDate(firstIncident.creationTime) || '--'),
     ).toBeInTheDocument();
     expect(withinRow.getByText(firstIncident.errorMessage)).toBeInTheDocument();
-
-    expect(
-      withinRow.getByRole('link', {
-        description: /view root cause instance/i,
-      }),
-    ).toBeInTheDocument();
     expect(
       withinRow.queryByRole('button', {name: 'Retry Incident'}),
     ).not.toBeInTheDocument();
 
     withinRow = within(
-      screen.getByRole('row', {
-        name: new RegExp(secondIncident.errorType.name),
-      }),
+      screen.getByRole('row', {name: new RegExp(secondIncidentErrorName)}),
     );
-    expect(
-      withinRow.getByText(secondIncident.errorType.name),
-    ).toBeInTheDocument();
-    expect(withinRow.getByText(secondIncident.flowNodeId)).toBeInTheDocument();
-    expect(withinRow.getByText(secondIncident.jobId!)).toBeInTheDocument();
+
+    expect(withinRow.getByText(secondIncidentErrorName)).toBeInTheDocument();
+    expect(withinRow.getByText(secondIncident.elementName)).toBeInTheDocument();
+    expect(withinRow.getByText(secondIncident.jobKey!)).toBeInTheDocument();
     expect(
       withinRow.getByText(formatDate(secondIncident.creationTime) || '--'),
     ).toBeInTheDocument();
@@ -170,61 +181,110 @@ describe('IncidentsTable', () => {
     expect(
       withinRow.queryByRole('button', {name: 'Retry Incident'}),
     ).not.toBeInTheDocument();
-    expect(
-      withinRow.queryByRole('link', {
-        description: /view root cause instance/i,
-      }),
-    ).not.toBeInTheDocument();
   });
 
-  it('should display -- for jobId', () => {
-    mockFetchProcessDefinitionXml().withSuccess('');
-    const incidentMock = {...firstIncident, jobId: null};
+  it('should display -- for jobKey', () => {
+    const incidentMock = {...firstIncident, jobKey: ''};
     const incidents = [incidentMock];
 
-    incidentsStore.setIncidents({...incidentsMock, incidents, count: 1});
-
-    render(<IncidentsTable />, {wrapper: Wrapper});
+    render(
+      <IncidentsTable
+        state="content"
+        processInstanceKey="1"
+        incidents={incidents}
+      />,
+      {wrapper: Wrapper},
+    );
 
     let withinFirstRow = within(
-      screen.getByRole('row', {
-        name: new RegExp(incidentMock.errorType.name),
-      }),
+      screen.getByRole('row', {name: new RegExp(firstIncidentErrorName)}),
     );
 
     expect(withinFirstRow.getByText('--')).toBeInTheDocument();
   });
 
-  it('should show a more button for long error messages', () => {
-    mockFetchProcessDefinitionXml().withSuccess('');
-    incidentsStore.setIncidents(incidentsMock);
-    render(<IncidentsTable />, {wrapper: Wrapper});
-    let withinFirstRow = within(
-      screen.getByRole('row', {
-        name: new RegExp(firstIncident.errorType.name),
+  it('should provide a link for incidents in child process instances', () => {
+    render(
+      <IncidentsTable
+        state="content"
+        processInstanceKey="7"
+        incidents={incidentsMock}
+      />,
+      {wrapper: Wrapper},
+    );
+    let withinRow = within(
+      screen.getByRole('row', {name: new RegExp(firstIncidentErrorName)}),
+    );
+
+    expect(
+      withinRow.getByRole('link', {
+        name: `${firstIncident.elementId} - ${firstIncident.processDefinitionName} - ${firstIncident.processInstanceKey}`,
+        description: `View root cause instance ${firstIncident.processDefinitionName} - ${firstIncident.processInstanceKey}`,
       }),
+    ).toBeInTheDocument();
+  });
+
+  it('should provide a retry operation for incidents in the open process instance', async () => {
+    render(
+      <IncidentsTable
+        state="content"
+        processInstanceKey="1"
+        incidents={[
+          {...firstIncident, processInstanceKey: '1'},
+          {...secondIncident, processInstanceKey: '2'},
+        ]}
+      />,
+      {wrapper: Wrapper},
+    );
+    let firstIncidentRow = within(
+      screen.getByRole('row', {name: new RegExp(firstIncidentErrorName)}),
+    );
+    let secondIncidentRow = within(
+      screen.getByRole('row', {name: new RegExp(secondIncidentErrorName)}),
+    );
+
+    expect(
+      await firstIncidentRow.findByRole('button', {name: 'Retry Incident'}),
+    ).toBeInTheDocument();
+    expect(
+      secondIncidentRow.queryByRole('button', {name: 'Retry Incident'}),
+    ).not.toBeInTheDocument();
+  });
+
+  it('should show a more button for long error messages', () => {
+    render(
+      <IncidentsTable
+        state="content"
+        processInstanceKey="1"
+        incidents={incidentsMock}
+      />,
+      {wrapper: Wrapper},
+    );
+    let withinFirstRow = within(
+      screen.getByRole('row', {name: new RegExp(firstIncidentErrorName)}),
     );
 
     expect(withinFirstRow.queryByText('More')).not.toBeInTheDocument();
 
     let withinSecondRow = within(
-      screen.getByRole('row', {
-        name: new RegExp(secondIncident.errorType.name),
-      }),
+      screen.getByRole('row', {name: new RegExp(secondIncidentErrorName)}),
     );
 
     expect(withinSecondRow.getByText('More')).toBeInTheDocument();
   });
 
   it('should open an modal when clicking on the more button', async () => {
-    incidentsStore.setIncidents(incidentsMock);
-    mockFetchProcessDefinitionXml().withSuccess('');
-    const {user} = render(<IncidentsTable />, {wrapper: Wrapper});
+    const {user} = render(
+      <IncidentsTable
+        state="content"
+        processInstanceKey="1"
+        incidents={incidentsMock}
+      />,
+      {wrapper: Wrapper},
+    );
 
     let withinSecondRow = within(
-      screen.getByRole('row', {
-        name: new RegExp(secondIncident.errorType.name),
-      }),
+      screen.getByRole('row', {name: new RegExp(secondIncidentErrorName)}),
     );
 
     expect(withinSecondRow.getByText('More')).toBeInTheDocument();
@@ -239,7 +299,7 @@ describe('IncidentsTable', () => {
       await within(modal).findByTestId('monaco-editor'),
     ).toBeInTheDocument();
     expect(
-      within(modal).getByText(`Flow Node "${secondIncident.flowNodeId}" Error`),
+      within(modal).getByText(`Element "${secondIncident.elementName}" Error`),
     ).toBeInTheDocument();
   });
 });
