+++ operate/client/src/App/ProcessInstance/IncidentsWrapper/IncidentsTable/tests/mocks.tsx
@@ -7,15 +7,15 @@
  */
 
 import {Route, MemoryRouter, Routes} from 'react-router-dom';
-import {createIncident} from 'modules/testUtils';
+import {createEnhancedIncidentV2} from 'modules/testUtils';
 import {useEffect} from 'react';
 import {authenticationStore} from 'modules/stores/authentication';
 import {incidentsStore} from 'modules/stores/incidents';
 import {flowNodeSelectionStore} from 'modules/stores/flowNodeSelection';
 import {Paths} from 'modules/Routes';
-import {ProcessDefinitionKeyContext} from 'App/Processes/ListView/processDefinitionKeyContext';
-import {QueryClientProvider} from '@tanstack/react-query';
 import {getMockQueryClient} from 'modules/react-query/mockQueryClient';
+import {QueryClientProvider} from '@tanstack/react-query';
+import {ProcessDefinitionKeyContext} from 'App/Processes/ListView/processDefinitionKeyContext';
 
 const Wrapper: React.FC<{children?: React.ReactNode}> = ({children}) => {
   useEffect(() => {
@@ -43,30 +43,23 @@ const shortError = 'No data found for query $.orderId.';
 const longError =
   'Cannot compare values of different types: INTEGER and BOOLEAN';
 
-const firstIncident = createIncident({
-  errorType: {name: 'Error A', id: 'ERROR_A'},
+const firstIncident = createEnhancedIncidentV2({
+  errorType: 'IO_MAPPING_ERROR',
+  processInstanceKey: '1',
   errorMessage: shortError,
-  flowNodeId: 'StartEvent_1',
-  flowNodeInstanceId: '18239123812938',
-  rootCauseInstance: {
-    instanceId: '111111111111111111',
-    processDefinitionId: 'calledInstance',
-    processDefinitionName: 'Called Instance',
-  },
+  elementId: 'StartEvent_1',
+  elementInstanceKey: '18239123812938',
+  processDefinitionId: 'calledInstance',
 });
 
-const secondIncident = createIncident({
-  errorType: {name: 'Error B', id: 'ERROR_A'},
+const secondIncident = createEnhancedIncidentV2({
+  errorType: 'CALLED_DECISION_ERROR',
+  processInstanceKey: '1',
   errorMessage: longError,
-  flowNodeId: 'Event_1db567d',
-  flowNodeInstanceId: id,
+  elementId: 'Event_1db567d',
+  elementInstanceKey: id,
 });
 
-const incidentsMock = {
-  incidents: [firstIncident, secondIncident],
-  count: 2,
-  errorTypes: [],
-  flowNodes: [],
-};
+const incidentsMock = [firstIncident, secondIncident];
 
-export {Wrapper, incidentsMock, firstIncident, secondIncident, shortError};
+export {Wrapper, incidentsMock, firstIncident, secondIncident};
