+++ operate/client/src/App/ProcessInstance/IncidentsWrapper/IncidentsTable/tests/selection.test.tsx
@@ -6,48 +6,61 @@
  * except in compliance with the Camunda License 1.0.
  */
 
-import {IncidentsTable} from '../index';
-import {createIncident} from 'modules/testUtils';
+import {IncidentsTable} from '..';
+import {createInstance, createProcessInstance} from 'modules/testUtils';
 import {render, screen} from 'modules/testing-library';
-import {incidentsStore} from 'modules/stores/incidents';
-import {flowNodeSelectionStore} from 'modules/stores/flowNodeSelection';
-import {Wrapper, incidentsMock, firstIncident} from './mocks';
-import {mockFetchProcessDefinitionXml} from 'modules/mocks/api/v2/processDefinitions/fetchProcessDefinitionXml';
+import {Wrapper, firstIncident, secondIncident} from './mocks';
+import * as selectionUtils from 'modules/utils/flowNodeSelection';
+import {mockFetchProcessInstance} from 'modules/mocks/api/processInstances/fetchProcessInstance';
+import {mockFetchProcessInstance as mockFetchProcessInstanceV2} from 'modules/mocks/api/v2/processInstances/fetchProcessInstance';
+
+// TanStack query is still fetching the data when the test executes, so "useRootNode"
+// falls back to undefined for the flowNodeInstanceId.
+const rootNode = {flowNodeInstanceId: undefined, isMultiInstance: false};
 
 describe('Selection', () => {
-  it('should deselect selected incident', async () => {
-    mockFetchProcessDefinitionXml().withSuccess('');
-    incidentsStore.setIncidents({
-      ...incidentsMock,
-      incidents: [firstIncident],
-      count: 1,
-    });
-    flowNodeSelectionStore.selectFlowNode({
-      flowNodeId: firstIncident.flowNodeId,
-      isMultiInstance: false,
-    });
+  beforeEach(() => {
+    mockFetchProcessInstance().withSuccess(createInstance());
+    mockFetchProcessInstanceV2().withSuccess(
+      createProcessInstance({
+        hasIncident: true,
+      }),
+    );
+  });
 
-    const {user} = render(<IncidentsTable />, {wrapper: Wrapper});
+  it('should deselect selected incident', async () => {
+    const spy = vi.spyOn(selectionUtils, 'clearSelection');
+
+    const {user} = render(
+      <IncidentsTable
+        state="content"
+        processInstanceKey="1"
+        incidents={[{...firstIncident, isSelected: true}]}
+      />,
+      {wrapper: Wrapper},
+    );
     expect(screen.getByRole('row', {selected: true})).toBeInTheDocument();
 
     await user.click(screen.getByRole('row', {selected: true}));
-    expect(screen.getByRole('row', {selected: false})).toBeInTheDocument();
+
+    expect(spy).toHaveBeenCalledTimes(1);
+    expect(spy).toHaveBeenCalledWith(rootNode);
   });
 
   it('should select single incident when multiple incidents are selected', async () => {
-    mockFetchProcessDefinitionXml().withSuccess('');
-    const incidents = [
-      createIncident({flowNodeId: 'myTask'}),
-      createIncident({flowNodeId: 'myTask'}),
-    ];
+    const spy = vi.spyOn(selectionUtils, 'selectFlowNode');
 
-    incidentsStore.setIncidents({...incidentsMock, incidents});
-    flowNodeSelectionStore.selectFlowNode({
-      flowNodeId: 'myTask',
-      isMultiInstance: false,
-    });
-
-    const {user} = render(<IncidentsTable />, {wrapper: Wrapper});
+    const {user} = render(
+      <IncidentsTable
+        state="content"
+        processInstanceKey="1"
+        incidents={[
+          {...firstIncident, isSelected: true, errorType: 'CONDITION_ERROR'},
+          {...secondIncident, isSelected: true, errorType: 'CONDITION_ERROR'},
+        ]}
+      />,
+      {wrapper: Wrapper},
+    );
     expect(screen.getAllByRole('row', {selected: true})).toHaveLength(2);
 
     const [firstRow] = screen.getAllByRole('row', {
@@ -55,19 +68,13 @@ describe('Selection', () => {
     });
 
     expect(firstRow).toBeInTheDocument();
-    await user.click(firstRow!);
+    await user.click(firstRow);
 
-    expect(
-      screen.getByRole('row', {
-        name: /condition error/i,
-        selected: true,
-      }),
-    ).toBeInTheDocument();
-    expect(
-      screen.getByRole('row', {
-        name: /condition error/i,
-        selected: false,
-      }),
-    ).toBeInTheDocument();
+    expect(spy).toHaveBeenCalledTimes(1);
+    expect(spy).toHaveBeenCalledWith(rootNode, {
+      flowNodeId: firstIncident.elementId,
+      flowNodeInstanceId: firstIncident.elementInstanceKey,
+      isMultiInstance: false,
+    });
   });
 });
