+++ operate/client/src/App/ProcessInstance/IncidentsWrapper/index.tsx
@@ -7,31 +7,19 @@
  */
 
 import {IncidentsOverlay} from './IncidentsOverlay';
-import {incidentsStore} from 'modules/stores/incidents';
 import {observer} from 'mobx-react';
 import {Transition} from './styled';
 import {IncidentsFilter} from './IncidentsFilter';
-import {IncidentsTable} from './IncidentsTable';
-import {IncidentsTable as IncidentsTableV2} from './IncidentsTable/v2';
+import {IncidentsTable as IncidentsTableV2} from './IncidentsTable';
 import {PanelHeader} from 'modules/components/PanelHeader';
-import {
-  getFilteredIncidents,
-  getIncidentsSearchFilter,
-  init,
-} from 'modules/utils/incidents';
-import {
-  useIncidents,
-  useEnhancedIncidents,
-  useIncidentsSort,
-} from 'modules/hooks/incidents';
+import {getIncidentsSearchFilter} from 'modules/utils/incidents';
+import {useEnhancedIncidents, useIncidentsSort} from 'modules/hooks/incidents';
 import {
   type Incident,
   type ProcessInstance,
   type QueryIncidentsResponseBody,
 } from '@camunda/camunda-api-zod-schemas/8.8';
 import type {InfiniteData, UseInfiniteQueryResult} from '@tanstack/react-query';
-import {useEffect} from 'react';
-import {IS_INCIDENTS_PANEL_V2} from 'modules/feature-flags';
 import {isInstanceRunning} from 'modules/utils/instance';
 import {modificationsStore} from 'modules/stores/modifications';
 import {useGetIncidentsByProcessInstancePaginated} from 'modules/queries/incidents/useGetIncidentsByProcessInstancePaginated';
@@ -46,62 +34,6 @@ type Props = {
 };
 
 const IncidentsWrapper: React.FC<Props> = observer(
-  ({setIsInTransition, processInstance}) => {
-    if (IS_INCIDENTS_PANEL_V2) {
-      // Having a condition before hooks is usually not allowed but works this time,
-      // because the condition is static during runtime.
-      return (
-        <IncidentsWrapperV2
-          setIsInTransition={setIsInTransition}
-          processInstance={processInstance}
-        />
-      );
-    }
-
-    const incidents = useIncidents();
-    const filteredIncidents = getFilteredIncidents(incidents);
-
-    useEffect(() => {
-      init(processInstance);
-
-      return () => {
-        incidentsStore.reset();
-      };
-    }, [processInstance]);
-
-    if (incidentsStore.incidentsCount === 0) {
-      return null;
-    }
-
-    return (
-      <>
-        <Transition
-          in={incidentsStore.state.isIncidentBarOpen}
-          onEnter={() => setIsInTransition(true)}
-          onEntered={() => setIsInTransition(false)}
-          onExit={() => setIsInTransition(true)}
-          onExited={() => setIsInTransition(false)}
-          mountOnEnter
-          unmountOnExit
-          timeout={400}
-        >
-          <IncidentsOverlay>
-            <PanelHeader
-              title="Incidents View"
-              count={filteredIncidents.length}
-              size="sm"
-            >
-              <IncidentsFilter />
-            </PanelHeader>
-            <IncidentsTable />
-          </IncidentsOverlay>
-        </Transition>
-      </>
-    );
-  },
-);
-
-const IncidentsWrapperV2: React.FC<Props> = observer(
   ({setIsInTransition, processInstance}) => {
     const enablePeriodicRefetch =
       isInstanceRunning(processInstance) &&
@@ -308,4 +240,4 @@ function computeDisplayStateFromQueryResult(
   }
 }
 
-export {IncidentsWrapper, IncidentsWrapperV2};
+export {IncidentsWrapper};
