+++ operate/client/src/App/ProcessInstance/TopPanel/MetadataPopover/Incidents/index.tsx
@@ -11,7 +11,6 @@ import {incidentsStore} from 'modules/stores/incidents';
 import {useGetIncidentsByElementInstance} from 'modules/queries/incidents/useGetIncidentsByElementInstance';
 import {MultiIncidents} from './multiIncidents';
 import {SingleIncident} from './singleIncident';
-import {IS_INCIDENTS_PANEL_V2} from 'modules/feature-flags';
 import {incidentsPanelStore} from 'modules/stores/incidentsPanel';
 import {Divider} from '../styled';
 
@@ -46,12 +45,11 @@ const Incidents: React.FC<Props> = ({
           <SingleIncident
             incident={singleIncident}
             onButtonClick={() => {
-              if (IS_INCIDENTS_PANEL_V2) {
-                return incidentsPanelStore.showIncidentsForElementInstance(
-                  elementInstanceKey,
-                  elementName,
-                );
-              }
+              incidentsPanelStore.showIncidentsForElementInstance(
+                elementInstanceKey,
+                elementName,
+              );
+
               incidentsStore.clearSelection();
               incidentsStore.toggleFlowNodeSelection(elementId);
               incidentsStore.toggleErrorTypeSelection(singleIncident.errorType);
@@ -65,12 +63,11 @@ const Incidents: React.FC<Props> = ({
           <MultiIncidents
             count={totalIncidents}
             onButtonClick={() => {
-              if (IS_INCIDENTS_PANEL_V2) {
-                return incidentsPanelStore.showIncidentsForElementInstance(
-                  elementInstanceKey,
-                  elementName,
-                );
-              }
+              incidentsPanelStore.showIncidentsForElementInstance(
+                elementInstanceKey,
+                elementName,
+              );
+
               incidentsStore.clearSelection();
               incidentsStore.toggleFlowNodeSelection(elementId);
               incidentsStore.setIncidentBarOpen(true);
