+++ operate/client/src/modules/components/IncidentOperation/index.tsx
@@ -6,13 +6,9 @@
  * except in compliance with the Camunda License 1.0.
  */
 
-import React, {useState} from 'react';
-
-import {type ErrorHandler, operationsStore} from 'modules/stores/operations';
+import React from 'react';
 import {OperationItems} from 'modules/components/OperationItems';
 import {OperationItem} from 'modules/components/OperationItem';
-import {observer} from 'mobx-react';
-
 import {tracking} from 'modules/tracking';
 import {InlineLoading} from '@carbon/react';
 import {Container} from './styled';
@@ -20,68 +16,12 @@ import {notificationsStore} from 'modules/stores/notifications';
 import {handleOperationError} from 'modules/utils/notifications';
 import {useResolveIncident} from 'modules/mutations/incidents/useResolveIncident';
 
-type Props = {
-  incidentKey: string;
-  instanceId: string;
-  showSpinner?: boolean;
-};
-
-const IncidentOperation: React.FC<Props> = observer(
-  ({instanceId, incidentKey, showSpinner}) => {
-    const [hasActiveOperation, setHasActiveOperation] = useState(false);
-
-    const handleError: ErrorHandler = ({statusCode}) => {
-      setHasActiveOperation(false);
-      handleOperationError(statusCode);
-    };
-
-    const handleOnClick = async (e: React.MouseEvent<HTMLButtonElement>) => {
-      e.stopPropagation();
-      setHasActiveOperation(true);
-
-      operationsStore.applyOperation({
-        instanceId,
-        payload: {
-          operationType: 'RESOLVE_INCIDENT',
-          incidentId: incidentKey,
-        },
-        onError: handleError,
-        onSuccess: () => {
-          tracking.track({
-            eventName: 'single-operation',
-            operationType: 'RESOLVE_INCIDENT',
-            source: 'incident-table',
-          });
-        },
-      });
-    };
-
-    return (
-      <Container orientation="horizontal">
-        {(hasActiveOperation || showSpinner) && (
-          <InlineLoading data-testid="operation-spinner" />
-        )}
-        <OperationItems>
-          <OperationItem
-            type="RESOLVE_INCIDENT"
-            onClick={handleOnClick}
-            data-testid="retry-incident"
-            title="Retry Incident"
-            disabled={hasActiveOperation || showSpinner}
-            size="sm"
-          />
-        </OperationItems>
-      </Container>
-    );
-  },
-);
-
 type IncidentOperationProps = {
   incidentKey: string;
   jobKey?: string;
 };
 
-const IncidentOperationV2: React.FC<IncidentOperationProps> = (props) => {
+const IncidentOperation: React.FC<IncidentOperationProps> = (props) => {
   const {isPending, mutate} = useResolveIncident(
     props.incidentKey,
     props.jobKey,
@@ -126,4 +66,4 @@ const IncidentOperationV2: React.FC<IncidentOperationProps> = (props) => {
   );
 };
 
-export {IncidentOperation, IncidentOperationV2};
+export {IncidentOperation};
