+++ operate/client/src/modules/feature-flags.ts
@@ -7,13 +7,11 @@
  */
 
 const IS_ADD_TOKEN_WITH_ANCESTOR_KEY_SUPPORTED = false;
-const IS_INCIDENTS_PANEL_V2 = false;
 const IS_PROCESSES_PAGE_V2 = false;
 const IS_ELEMENT_INSTANCE_HISTORY_V2 = false;
 
 export {
   IS_ADD_TOKEN_WITH_ANCESTOR_KEY_SUPPORTED,
-  IS_INCIDENTS_PANEL_V2,
   IS_PROCESSES_PAGE_V2,
   IS_ELEMENT_INSTANCE_HISTORY_V2,
 };
