+++ operate/client/src/modules/utils/incidents.ts
@@ -12,7 +12,7 @@ import type {
   QueryIncidentsRequestBody,
 } from '@camunda/camunda-api-zod-schemas/8.8';
 import {autorun} from 'mobx';
-import {incidentsStore, type Incident} from 'modules/stores/incidents';
+import {incidentsStore} from 'modules/stores/incidents';
 import {isInstanceRunning} from './instance';
 import type {EnhancedIncident} from 'modules/hooks/incidents';
 
@@ -81,45 +81,6 @@ const startPolling = async (
 };
 
 const isSingleIncidentSelected = (
-  incidents: Incident[],
-  flowNodeInstanceId: string,
-) => {
-  const selectedInstances = incidents.filter((incident) => incident.isSelected);
-
-  return (
-    selectedInstances.length === 1 &&
-    selectedInstances[0]?.flowNodeInstanceId === flowNodeInstanceId
-  );
-};
-
-const getFilteredIncidents = (incidents: Incident[]) => {
-  const {selectedFlowNodes, selectedErrorTypes} = incidentsStore.state;
-
-  const hasSelectedFlowNodes = selectedFlowNodes.length > 0;
-  const hasSelectedErrorTypes = selectedErrorTypes.length > 0;
-
-  if (!hasSelectedFlowNodes && !hasSelectedErrorTypes) {
-    return incidents;
-  }
-
-  return incidents.filter((incident) => {
-    if (hasSelectedErrorTypes && hasSelectedFlowNodes) {
-      return (
-        selectedErrorTypes.includes(incident.errorType.id) &&
-        selectedFlowNodes.includes(incident.flowNodeId)
-      );
-    }
-    if (hasSelectedErrorTypes) {
-      return selectedErrorTypes.includes(incident.errorType.id);
-    }
-    if (hasSelectedFlowNodes) {
-      return selectedFlowNodes.includes(incident.flowNodeId);
-    }
-    return [];
-  });
-};
-
-const isSingleIncidentSelectedV2 = (
   incidents: EnhancedIncident[],
   elementInstanceKey: string,
 ) => {
@@ -145,7 +106,5 @@ export {
   init,
   startPolling,
   isSingleIncidentSelected,
-  isSingleIncidentSelectedV2,
-  getFilteredIncidents,
   getIncidentsSearchFilter,
 };
